// src/utils/OpenAI/OpenAI.types.ts
import { OpenAIChatModels } from "./OpenAI.constants";

export type JulepAIChatMessageRole = "system" | "assistant" | "user";

export interface OpenAIChatMessage {
  id?: number;
  role: JulepAIChatMessageRole;
  content: string;
  name?: string;
}

export interface OpenAISystemMessage {
  role: "system";
  name: string;
  content: string;
}

export interface OpenAIConfig {
  model: keyof typeof OpenAIChatModels;
  temperature?: number;
  max_tokens?: number;
  top_p?: number;
  frequency_penalty?: number;
  presence_penalty?: number;
  ignore_eos?: boolean;
  stop?: string[];
  skip_special_tokens?: boolean;
}

export interface OpenAIModel {
  id: string;
  name: string;
  maxLimit: number;
}